+++ ui/src/features/ui/index.ts
@@ -5,6 +5,7 @@ export * from './creatable-select'
 export * from './delete-button'
 export * from './description'
 export * from './drawer'
+export * from './endpoint'
 export * from './field'
 export * from './form'
 export * from './group-select'
